Extract ErrorDetails class and add Errors prop
